+++ package.json
@@ -14,8 +14,7 @@
     "request": "^2.88.2"
   },
   "devDependencies": {
-    "@pact-foundation/pact": "^9.17.3",
-    "@pact-foundation/pact-node": "^10.17.2",
+    "@pact-foundation/pact": "^10.1.4",
     "dotenv": "^8.2.0",
     "eslint": "^8.12.0",
     "jest": "^27.5.1"
